Add %yyinit {} to customize scanner initialization
This is equivalent to the YY_USER_INIT macro in flex() or
the %begin {} code block in RE-flex.

- recognize the %yyinit {} code block and record the code
  in a specific temporary file
- update the template expander to replace %yyinit in the
  template by the user code found for %yyinit {} directive

It can be used to put the scanner in a specific state at the
beginning of parsing a new input.  Example:

%s START
%yyinit {
   ENTER(START);
}

<START>build { ENTER (INITIAL); return T_BUILD; }

Here, the scanner is initialized in the START condition instead
of INITIAL and the <START>build rule can be matched.  Once that
rule is matched, the scanner switches to the INITIAL condition
(which is its default) and the build token is no longer recognized.